+++ appdaemon/utility_loop.py
@@ -77,7 +77,10 @@ class Utility:
                 # Register set_state services
                 #
             
-                self.AD.services.register_service(ns, "state", "set", self.AD.state.state_services)
+                # only default, rules or it belongs to a local plugin. Don't allow for admin/appdaemon/global namespaces
+
+                if ns in ["default", "rules"] or ns in self.AD.plugins.plugin_objs or ns in self.AD.namespaces: 
+                    self.AD.services.register_service(ns, "state", "set", self.AD.state.state_services)
                 
                 #
                 # Register fire_event services
@@ -91,6 +94,11 @@ class Utility:
             # Register run_sequence service
             #
             self.AD.services.register_service("rules", "sequence", "run", self.AD.sequences.run_sequence_service)
+            
+            #
+            # Register production_mode service
+            #
+            self.AD.services.register_service("appdaemon", "production_mode", "set", self.production_mode_service)
 
             #
             # Start the scheduler
@@ -203,3 +211,11 @@ class Utility:
         else:
             self.logger.info("AD Production Mode Deactivated")
         self.AD.production_mode = mode
+    
+    async def production_mode_service(self, ns, domain, service, kwargs):
+        if "mode" in kwargs:
+            mode = kwargs["mode"]
+            await self.set_production_mode(mode)
+        else:
+            self.logger.warning("'Mode' not specified in service call")
+