+++ docs/HISTORY.rst
@@ -8,10 +8,12 @@ Change Log
 
 - Added events for when an app is initialized or terminated
 - Added `event_fire` service call
+- Added `production_mode` service call
 
 **Fixes**
 
 - Fixed an issue, where when ``http`` is disabled in ``appdaemon.yaml``, AD is unable to start
+- Fixed issue where the user could potentially create entities in `admin`, `global` or `appdaemon` namespaces
 
 **Breaking Changes**